+++ R/peddown.R
@@ -6,19 +6,29 @@
 #'
 #' @param parents : Vector of individual ID(s), from which the new pedigree is being extracted.
 #'
+#' @param maxgen : (optional) a positive integer for the maximum number of generations to proceed. If no value is provided, there is no limitation on the maximum number of generations to proceed.
+#'
 #' @return newped : Extracted pedigree \code{data.frame}
 #'
 #' @examples
 #' ped = data.frame(ID=1:6, SIRE=c(0,0,1,3,1,4), DAM=c(0,0,2,2,2,5))
 #' peddown(ped, c(1,4))
+#' peddown(ped, 1, maxgen=1)
 #'
 #' @export
-peddown = function(ped, parents) {
+peddown = function(ped, parents, maxgen=c()) {
+   if(length(maxgen) >1) stop("Provide a positive integer for maxgen.")
+   if(length(maxgen)==1) {
+     if(maxgen!=round(maxgen) | maxgen<1) stop("Invalid maxgen")
+   }
    colnames(ped) = c("ID","SIRE","DAM")
    oldped = data.frame()
    newped = data.frame(ID=parents, SIRE=0, DAM=0)
    curr.parents = parents
-   while(nrow(oldped) < nrow(newped))
+   iter = 0
+   maxgen2 = maxgen
+   if(length(maxgen)==0) maxgen2 = 1
+   while(nrow(oldped) < nrow(newped) & iter < maxgen2)
    {
       oldped = newped
       tmp = ped[ped$SIRE %in% curr.parents | ped$DAM %in% curr.parents,]
@@ -33,6 +43,7 @@ peddown = function(ped, parents) {
          }
       }
       curr.parents = tmp$ID
+      if(length(maxgen)==1) iter = iter + 1
    }
    newped[!newped$SIRE %in% newped$ID,]$SIRE = 0
    newped[!newped$DAM  %in% newped$ID,]$DAM  = 0